🔀 refactor(Button): refactor Button component to support different variants of buttons
The Button component has been refactored to support different variants of buttons. Previously, it only had a single variant with a Plus icon. Now, it has two variants: "ADD" and "MORE".

- The "ADD" variant displays an Add icon.
- The "MORE" variant displays a More icon.

The variant can be specified using the `variant` prop. If no variant is specified, the default variant is "ADD".

This change improves the flexibility and reusability of the Button component, allowing it to be used in different contexts with different icons.