✨ feat(getUnitLabelByMeasureValue.ts): add function to get unit label based on measure value and unit type
The `getUnitLabelByMeasureValue` function is added to the `utils` directory. This function takes in a `measureValue` and a `unit` as parameters and returns the corresponding unit label based on the measure value and unit type. If the `measureValue` is equal to 1, the function returns the singular label for the unit. Otherwise, it returns the plural label for the unit.